Introduce `cache-directory` config option to speed up analysis
We use `ComposerRequireChecker` for our monolith with 10k+ classes and run it on GitHub Actions.
The average run time is 5 minutes and 30 seconds. And that's only `ComposerRequireChecker`. Together with the job setup it's even slower.

By setting the `cache-directory` in our configuration file we are able to cache the parsed PHP files
and reuse them for the next run.

Now, the checker runs in 2 minute and 30 seconds. Making the total difference 5:30 => 2:00 = 3 minutes faster.

Now there was already a discussion in #295 if this was needed or not and what is considered fast or not. For us, we spend a lot of time in optimizing our CI pipeline to let it finish within a few minutes. Shaving off 3 minutes is worth it for us. I can understand that you might not want to merge this in, and that's fine. I'll resort to my own fork then.